apps/api/pyproject.toml
[project]
name = "wishlist-api"
version = "0.1.0"
description = "Social wishlist backend API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn[standard]>=0.30.0,<1.0.0",
  "sqlalchemy[asyncio]>=2.0.32,<3.0.0",
  "asyncpg>=0.29.0,<1.0.0",
  "alembic>=1.13.2,<2.0.0",
  "pydantic-settings>=2.4.0,<3.0.0",
  "python-jose[cryptography]>=3.3.0,<4.0.0",
  "passlib>=1.7.4,<2.0.0",
  "email-validator>=2.2.0,<3.0.0",
  "authlib>=1.3.1,<2.0.0",
  "httpx>=0.27.0,<1.0.0",
  "beautifulsoup4>=4.12.3,<5.0.0",
  "python-multipart>=0.0.9,<1.0.0",
  "orjson>=3.10.7,<4.0.0",
  "itsdangerous>=2.2.0,<3.0.0",
  "greenlet>=3.0.3,<4.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.2,<9.0.0",
  "pytest-asyncio>=0.23.8,<1.0.0",
  "pytest-cov>=5.0.0,<6.0.0",
  "ruff>=0.6.2,<1.0.0",
  "mypy>=1.11.1,<2.0.0",
  "aiosqlite>=0.20.0,<1.0.0",
]

[build-system]
requires = ["setuptools>=72.1.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff]
line-length = 140
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "B"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = false
ignore_missing_imports = true
